+++ src/blas_helper.h
@@ -36,7 +36,7 @@ namespace libsdp{
 #define F_DAXPY daxpy_
 #define F_DCOPY dcopy_
 #define F_DSCAL dscal_
-#define F_DDOT ddot_
+#define F_DDOT  ddot_
 #define F_DNRM2 dnrm2_
 #define F_DGEMM dgemm_
 #define F_DGEMV dgemv_
@@ -45,26 +45,26 @@ namespace libsdp{
 extern "C" {
 
 
-extern void F_DGEMM(char*, char*, int*, int*, int*, double*, double*, int*, double*, int*, double*, double*, int*);
-extern void F_DAXPY(int *length, double *a, double *x, int *inc_x, double *y, int *inc_y);
-extern void F_DCOPY(int *length, double *x, int *inc_x, double *y, int *inc_y);
-extern void F_DSCAL(int *n, double *alpha, double *vec, int *inc);
-extern double F_DDOT(int *n, double *x, int *incx, double *y, int *incy);
-extern double F_DNRM2(int *n, double *x, int *incx);
-extern void F_DSYEV(char &JOBZ, char &UPLO, int &N, double *A, int &LDA, double *W, double *WORK,int &LWORK, int &INFO);
+extern void   F_DGEMM(char*, char*, long int*, long int*, long int*, double*, double*, long int*, double*, long int*, double*, double*, long int*);
+extern void   F_DAXPY(long int *length, double *a, double *x, long int *inc_x, double *y, long int *inc_y);
+extern void   F_DCOPY(long int *length, double *x, long int *inc_x, double *y, long int *inc_y);
+extern void   F_DSCAL(long int *n, double *alpha, double *vec, long int *inc);
+extern double F_DDOT(long int *n, double *x, long int *incx, double *y, long int *incy);
+extern double F_DNRM2(long int *n, double *x, long int *incx);
+extern void   F_DSYEV(char &JOBZ, char &UPLO, long int &N, double *A, long int &LDA, double *W, double *WORK,long int &LWORK, long int &INFO);
 
 }
 
-void C_DSCAL(size_t len, double alpha, double* vec, int inc);
-void C_DCOPY(size_t length, double* x, int inc_x, double* y, int inc_y);
-void C_DAXPY(size_t length, double a, double* x, int inc_x, double* y, int inc_y);
-double C_DDOT(size_t n, double* X, int inc_x, double* Y, int inc_y);
-double C_DNRM2(size_t n, double* X, int inc_x);
+void C_DSCAL(size_t len, double alpha, double* vec, long int inc);
+void C_DCOPY(size_t length, double* x, long int inc_x, double* y, long int inc_y);
+void C_DAXPY(size_t length, double a, double* x, long int inc_x, double* y, long int inc_y);
+double C_DDOT(size_t n, double* X, long int inc_x, double* Y, long int inc_y);
+double C_DNRM2(size_t n, double* X, long int inc_x);
 
 /**
  * diagonalize a real symmetric matrix
  */
-void Diagonalize(long int N, double *A, double *W);
+void Diagonalize(long long int N, double *A, double *W);
 
 
 } // end of namespace
